# swaggerV2.json
{
  "swagger": "2.0",
  "info": {
    "description": "RESTfull API baseada em node.js Com o objetivo da partilha de viagens e experiencias vividas, classificações e comentáriosentre utilizadores. Disponível em https://tppweb.herokuapp.com",
    "version": "2.0.0",
    "title": "Trips Together",
    "contact": {
      "email": "o.nosso.email@gmail.com"
    }
  },
...
  "produces": [
    "application/json"
  ],
  "host": "tppweb.herokuapp.com",
  "basePath": "/api/v2",
  "paths": {
...
    "/viagens/utilizador/{idUtilizador}": {
      "get": {
        "tags": [
          "Viagens"
        ],
        "summary": "Filtra todas as viagens por utilizador",
        "parameters": [
          {
            "name": "idUtilizador",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Viages obtidas com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/utilizadores": {
      "get": {
        "tags": [
          "Utilizadores"
        ],
        "summary": "Obter utilizadores",
        "description": "Obtem uma listagem com todos os utilizadores ou filtrados por nome",
        "parameters": [
          {
            "name": "filtro",
            "in": "query",
            "description": "Filtro a ser aplicado no campo nome do utilizador",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
    "/utilizadores/{id}": {
      "get": {
        "tags": [
          "Utilizadores"
        ],
        "summary": "Obter um utilizador",
        "description": "Obtem um utilizador atraves do seu ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/bucketlist/utilizador/{id}": {
      "get": {
        "tags": [
          "Bucket list"
        ],
        "summary": "Obter bucket list de viagens de um utilizador",
        "description": "Obtem todos as viagens existentes na bucket list de um utilizador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/comentarios/utilizador/{id}": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Obter os comentários de um utilizador",
        "description": "Obtem todos os comentários de um determinado utilizador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/notificacoes/utilizador/{id}": {
      "get": {
        "tags": [
          "Notificações"
        ],
        "summary": "Todas as notificações de um utilizador",
        "description": "Obtem todas as notificações de um utilizador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID utilizador",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Response": {
      "type": "object",
      "properties": {
        "ErroCorde": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        },
        "Error": {
          "type": "object"
        },
        "Data": {
          "type": "object"
        }
      }
    }
  }
}